+++ docs/index.d.ts
@@ -1074,6 +1074,10 @@ export interface MempoolTokenTransferTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -1117,6 +1121,10 @@ export interface MempoolSmartContractTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -1157,6 +1165,10 @@ export interface MempoolContractCallTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -1196,6 +1208,10 @@ export interface MempoolPoisonMicroblockTransaction {
     hex: string;
     repr: string;
   };
+  /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
   /**
    * Integer string (64-bit unsigned integer).
    */
@@ -1238,6 +1254,10 @@ export interface MempoolCoinbaseTransaction {
     hex: string;
     repr: string;
   };
+  /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
   /**
    * Integer string (64-bit unsigned integer).
    */
@@ -2050,6 +2070,10 @@ export interface TokenTransferTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Transaction fee as Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -2128,6 +2152,10 @@ export interface SmartContractTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Transaction fee as Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -2205,6 +2233,10 @@ export interface ContractCallTransaction {
      */
     repr: string;
   };
+  /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
   /**
    * Transaction fee as Integer string (64-bit unsigned integer).
    */
@@ -2297,6 +2329,10 @@ export interface PoisonMicroblockTransaction {
     repr: string;
   };
   /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
+  /**
    * Transaction fee as Integer string (64-bit unsigned integer).
    */
   fee_rate: string;
@@ -2369,6 +2405,10 @@ export interface CoinbaseTransaction {
      */
     repr: string;
   };
+  /**
+   * Used for ordering the transactions originating from and paying from an account. The nonce ensures that a transaction is processed at most once. The nonce counts the number of times an account's owner(s) have authorized a transaction. The first transaction from an account will have a nonce value equal to 0, the second will have a nonce value equal to 1, and so on.
+   */
+  nonce: number;
   /**
    * Transaction fee as Integer string (64-bit unsigned integer).
    */