+++ src/migrations/1591291822107_mempool_txs.ts
@@ -26,6 +26,10 @@ export async function up(pgm: MigrationBuilder): Promise<void> {
       type: 'bytea',
       notNull: true,
     },
+    nonce: {
+      type: 'integer',
+      notNull: true,
+    },
     fee_rate: {
       type: 'bigint',
       notNull: true,