+++ src/tests-rosetta-cli/validate-rosetta.ts
@@ -16,6 +16,7 @@ import {
   estimateContractFunctionCall,
   ClarityAbi,
   encodeClarityValue,
+  ChainID,
 } from '@stacks/transactions';
 import { StacksTestnet } from '@stacks/network';
 import * as BN from 'bn.js';
@@ -124,8 +125,8 @@ describe('Rosetta API', () => {
     await cycleMigrations();
     db = await PgDataStore.connect();
     client = await db.pool.connect();
-    eventServer = await startEventServer({ db });
-    api = await startApiServer(db);
+    eventServer = await startEventServer({ db, chainId: ChainID.Testnet });
+    api = await startApiServer(db, ChainID.Testnet);
 
     // remove previous outputs if any
     fs.rmdirSync('rosetta-output', { recursive: true });
@@ -198,7 +199,7 @@ describe('Rosetta API', () => {
   });
 
   afterAll(async () => {
-    await new Promise(resolve => eventServer.close(() => resolve()));
+    await new Promise<void>(resolve => eventServer.close(() => resolve()));
     await api.terminate();
     client.release();
     await db?.close();