+++ src/tests/teardown.ts
@@ -1,7 +1,7 @@
 export default async (): Promise<void> => {
   console.log('Jest - teardown..');
   const eventSocketServer: import('net').Server = (global as any).server;
-  await new Promise(resolve => {
+  await new Promise<void>(resolve => {
     eventSocketServer.close(() => {
       console.log('Jest - teardown done');
       resolve();